+++ packages/flutter/lib/painting.dart
@@ -21,6 +21,7 @@ export 'src/painting/alignment.dart';
 export 'src/painting/basic_types.dart';
 export 'src/painting/border_radius.dart';
 export 'src/painting/borders.dart';
+export 'src/painting/box_border.dart';
 export 'src/painting/box_decoration.dart';
 export 'src/painting/box_fit.dart';
 export 'src/painting/box_shadow.dart';
